Refactor semantic assembler into a modular structure
- Split `src/semantic/assembler.rs` into `src/semantic/assembler/` directory.
- Extracted data models to `model.rs`.
- Extracted errors to `errors.rs`.
- Moved core logic and tests to `mod.rs`.
- Updated `atlas.md` journal.
